AJAX requests for operations assign

+++ production/views/technologist_views.py
@@ -80,7 +80,6 @@ class SizeQuantityCreateView(View):
             'passport': passport
         })
     def post(self, request, passport_id):
-        print(request.headers)
         if request.headers.get('X-Requested-With') == 'XMLHttpRequest':
             form = SizeQuantityForm(request.POST)
             if form.is_valid():
@@ -124,46 +123,52 @@ def assign_operations(request, passport_id):
     operations = passport.order.model.operations.all()
     size_quantities = passport.size_quantities.all()
 
-    if request.method == 'POST':
-        with transaction.atomic():
-            errors = False
-            for key, input_value in request.POST.items():
-                if key.startswith('employee_') and input_value:
-                    _, operation_id, size_quantity_id = key.split('_')
-                    entries = [entry.strip() for entry in input_value.split(',')]
-                    total_quantity = passport.size_quantities.get(id=size_quantity_id).quantity
-                    
-                    # Reset quantities for existing assigned work to avoid duplication
-                    Work.objects.filter(operation_id=operation_id, size_quantity_id=size_quantity_id, passport=passport).delete()
+    if request.method == 'POST' and request.headers.get('X-Requested-With') == 'XMLHttpRequest':
+        # Decode the JSON data from the request
+        data = json.loads(request.body)
+        operation_id = data.get('operation_id')
+        size_quantity_id = data.get('size_quantity_id')
+        value = data.get('value')
+        try:
+            with transaction.atomic():
+                errors = False
 
-                    for entry in entries:
-                        if '(' in entry and ')' in entry:
-                            employee_id_input, quantity = entry.split('(')
-                            employee_id_input = employee_id_input.strip()
-                            quantity = int(quantity.strip(' )'))
-                        else:
-                            employee_id_input = entry
-                            quantity = total_quantity
+                entries = [entry.strip() for entry in value.split(',')]
+                total_quantity = passport.size_quantities.get(id=size_quantity_id).quantity
 
-                        if employee_id_input:
-                            employee_profile = UserProfile.objects.filter(employee_id=employee_id_input, type=UserProfile.EMPLOYEE).first()
-                            if not employee_profile:
-                                messages.error(request, f'Invalid employee ID: {employee_id_input}')
-                                errors = True
-                                continue
-                            
-                            work, created = Work.objects.get_or_create(
-                                operation_id=operation_id,
-                                size_quantity_id=size_quantity_id,
-                                passport=passport
-                            )
-                            AssignedWork.objects.create(
-                                work=work,
-                                employee=employee_profile,
-                                quantity=quantity
-                            )
-            if not errors:
-                return HttpResponseRedirect(request.path_info)
+                # Reset quantities for existing assigned work to avoid duplication
+                Work.objects.filter(operation_id=operation_id, size_quantity_id=size_quantity_id, passport=passport).delete()
+
+                for entry in entries:
+                    if '(' in entry and ')' in entry:
+                        employee_id_input, quantity = entry.split('(')
+                        employee_id_input = employee_id_input.strip()
+                        quantity = int(quantity.strip(' )'))
+                    else:
+                        employee_id_input = entry
+                        quantity = total_quantity
+
+                    if employee_id_input:
+                        employee_profile = UserProfile.objects.filter(employee_id=employee_id_input, type=UserProfile.EMPLOYEE).first()
+                        if not employee_profile:
+                            errors = True
+                            continue
+
+                        work, created = Work.objects.get_or_create(
+                            operation_id=operation_id,
+                            size_quantity_id=size_quantity_id,
+                            passport=passport
+                        )
+                        AssignedWork.objects.create(
+                            work=work,
+                            employee=employee_profile,
+                            quantity=quantity
+                        )
+                if not errors:
+                    return JsonResponse({'status': 'success'})
+
+        except Exception as e:
+            return JsonResponse({'status': 'error', 'message': str(e)}, status=400)
 
     work_by_op_and_size = {}
     for assigned_work in AssignedWork.objects.filter(work__passport=passport).select_related('employee', 'work__operation', 'work__size_quantity'):
@@ -173,8 +178,8 @@ def assign_operations(request, passport_id):
         else:
             work_by_op_and_size[key].append(assigned_work)
     return render(request, 'technologist/passports/assign_operations.html', {
-        'passport': passport, 
-        'operations': operations, 
+        'passport': passport,
+        'operations': operations,
         'size_quantities': size_quantities,
         'work_by_op_and_size': work_by_op_and_size
     })
